getting ready for release

+++ core/compiler/emit.h
@@ -66,7 +66,7 @@ namespace backend {
       out_stream.WriteByte(value);
     }
 
-    inline void WriteInt(long value, OutputStream& out_stream) {
+    inline void WriteInt(int32_t value, OutputStream& out_stream) {
       out_stream.WriteInt(value);
     }
